renames files and refactors code to match style guide.
