+++ src/Components/Upvote/Upvote.styles.jsx
@@ -1,0 +1,14 @@
+import Styled from "styled-components";
+
+export const UpvoteDiv = Styled.div`
+button {
+    width: 100px;
+    height: 1.6rem;
+    border-radius: 7px;
+    background-color: #3184ed;
+    font-weight: 500;
+    color: white;
+    border: unset;
+    cursor: pointer;
+}
+`;
