--- src/contexts/context.js
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const FeedContext = createContext();
